refactor. 마이페이지 컴포넌트 위치 이동 #33
마이페이지에서 사용하는 컴포넌트 my 폴더로 이동
