+++ src/components/Header.tsx
@@ -10,7 +10,7 @@ import useViewport from '../hooks/useViewport';
 import * as t from '../style/header.style';
 import GlobalModal from './common/GlobalModal';
 import MenuDrop from './common/MenuDrop';
-import HeaderAside from './HeaderAside';
+import HeaderAside from './my/HeaderAside';
 import SearchBar from './SearchBar';
 
 export default function Header() {
