+++ src/components/my/MyUserHeadSection.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import useViewport from '../hooks/useViewport';
-import * as t from '../style/myUserHeadSection.style';
-import GlobalModal from './common/GlobalModal';
-import MyUserInfoModal from './common/UserInfoModal';
+import useViewport from '../../hooks/useViewport';
+import * as t from '../../style/myUserHeadSection.style';
+import GlobalModal from '../common/GlobalModal';
+import MyUserInfoModal from '../common/UserInfoModal';
 
 interface IUser {
   u_Img: string;
